union mypagecontroller into one file and add seoul json
