+++ src/controller/mypage_controller.java
@@ -1,6 +1,7 @@
 package controller;
 
 import java.io.IOException;
+import java.time.LocalDate;
 import java.util.List;
 
 import javax.servlet.RequestDispatcher;
@@ -12,6 +13,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import mypageDao.NoticeDao;
 import mypageDto.NoticeDto;
+import result.result_dao.result_dao;
+import result.result_dto.result_dto;
 
 /**
  * Servlet implementation class Notice
@@ -29,6 +32,7 @@ public class mypage_controller extends HttpServlet {
 		System.out.println("[command: " + command + "]");
 		
 		NoticeDao dao = new NoticeDao();
+		result_dao resultdao = new result_dao();
 		
 		if(command.equals("main")) {
 			response.sendRedirect("main.jsp");
@@ -60,7 +64,37 @@ public class mypage_controller extends HttpServlet {
 		else if(command.equals("mypage_qna")) {
 			
 		}
-		
+		/*상품교환 페이지 이동*/
+		else if(command.equals("mypage_changestar")){
+			response.sendRedirect("mypage/mypage_changestar.jsp");
+		}
+		/*성적 확인 페이지 이동*/
+		else if(command.equals("mypage_checkscore")){
+			result_dto month_1 = resultdao.select_month();
+			result_dto month_2 = resultdao.select_month_1();
+			result_dto month_3 = resultdao.select_month_2();
+
+			LocalDate now = LocalDate.now();
+
+			int month = now.getMonthValue();
+
+			if(month_1.getMonth() ==  month){
+				request.setAttribute("month_1", month_1);
+			}
+			if(month_2.getMonth() == (month-1)){
+				request.setAttribute("month_2", month_2);
+			}
+			if(month_3.getMonth() == (month-2)){
+				request.setAttribute("month_3", month_3);
+			}
+
+			dispatch("mypage/mypage_checkscore.jsp", request, response);
+		}
+	}
+
+	private void dispatch(String url, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		RequestDispatcher dispatch = request.getRequestDispatcher(url);
+		dispatch.forward(request, response);
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
